# docs/backend/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Diabetics Platform API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/nutrition/ingredients": {
      "get": {
        "tags": [
          "Nutrition Engine"
        ],
        "summary": "Search Ingredients",
        "description": "Búsqueda por nombre de ingredientes.",
        "operationId": "search_ingredients_api_v1_nutrition_ingredients_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 2,
              "description": "Nombre del alimento a buscar",
              "title": "Q"
            },
            "description": "Nombre del alimento a buscar"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IngredientResponse"
                  },
                  "title": "Response Search Ingredients Api V1 Nutrition Ingredients Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Nutrition Engine"
        ],
        "summary": "Create Ingredient",
        "description": "Crea un nuevo ingrediente en la base de datos.",
        "operationId": "create_ingredient_api_v1_nutrition_ingredients_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngredientCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngredientResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/nutrition/bolus/calculate": {
      "post": {
        "tags": [
          "Nutrition Engine"
        ],
        "summary": "Calculate Bolus",
        "description": "Calcula el bolus de insulina basado en carbohidratos, glucemia actual y objetivos.",
        "operationId": "calculate_bolus_api_v1_nutrition_bolus_calculate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BolusCalcRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BolusCalcResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/nutrition/meals": {
      "post": {
        "tags": [
          "Nutrition Engine"
        ],
        "summary": "Log Meal",
        "description": "Registra una ingesta en el historial del paciente, encriptando notas (PHI).",
        "operationId": "log_meal_api_v1_nutrition_meals_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogMealRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MealLogResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/nutrition/meals/history": {
      "get": {
        "tags": [
          "Nutrition Engine"
        ],
        "summary": "Get Meal History",
        "description": "Devuelve el historial de comidas registradas para un paciente, más reciente primero.",
        "operationId": "get_meal_history_api_v1_nutrition_meals_history_get",
        "parameters": [
          {
            "name": "patient_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Patient Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtro fecha inicio (ISO 8601)",
              "title": "Start Date"
            },
            "description": "Filtro fecha inicio (ISO 8601)"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtro fecha fin (ISO 8601)",
              "title": "End Date"
            },
            "description": "Filtro fecha fin (ISO 8601)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MealLogResponse"
                  },
                  "title": "Response Get Meal History Api V1 Nutrition Meals History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BolusCalcRequest": {
        "properties": {
          "current_glucose": {
            "type": "number",
            "title": "Current Glucose"
          },
          "target_glucose": {
            "type": "number",
            "title": "Target Glucose"
          },
          "ingredients": {
            "items": {
              "$ref": "#/components/schemas/IngredientInput"
            },
            "type": "array",
            "title": "Ingredients"
          },
          "icr": {
            "type": "number",
            "title": "Icr",
            "default": 10.0
          },
          "isf": {
            "type": "number",
            "title": "Isf",
            "default": 50.0
          }
        },
        "type": "object",
        "required": [
          "current_glucose",
          "target_glucose",
          "ingredients"
        ],
        "title": "BolusCalcRequest"
      },
      "BolusCalcResponse": {
        "properties": {
          "total_carbs_grams": {
            "type": "number",
            "title": "Total Carbs Grams"
          },
          "recommended_bolus_units": {
            "type": "number",
            "title": "Recommended Bolus Units"
          }
        },
        "type": "object",
        "required": [
          "total_carbs_grams",
          "recommended_bolus_units"
        ],
        "title": "BolusCalcResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IngredientCreateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "glycemic_index": {
            "type": "integer",
            "title": "Glycemic Index"
          },
          "carbs_per_100g": {
            "type": "number",
            "title": "Carbs Per 100G"
          },
          "fiber_per_100g": {
            "type": "number",
            "title": "Fiber Per 100G",
            "default": 0.0
          }
        },
        "type": "object",
        "required": [
          "name",
          "glycemic_index",
          "carbs_per_100g"
        ],
        "title": "IngredientCreateRequest"
      },
      "IngredientInput": {
        "properties": {
          "ingredient_id": {
            "type": "string",
            "format": "uuid",
            "title": "Ingredient Id"
          },
          "weight_grams": {
            "type": "number",
            "title": "Weight Grams"
          }
        },
        "type": "object",
        "required": [
          "ingredient_id",
          "weight_grams"
        ],
        "title": "IngredientInput"
      },
      "IngredientResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "glycemic_index": {
            "type": "integer",
            "title": "Glycemic Index"
          },
          "carbs": {
            "type": "number",
            "title": "Carbs"
          },
          "fiber_per_100g": {
            "type": "number",
            "title": "Fiber Per 100G"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "glycemic_index",
          "carbs",
          "fiber_per_100g"
        ],
        "title": "IngredientResponse"
      },
      "LogMealRequest": {
        "properties": {
          "patient_id": {
            "type": "string",
            "format": "uuid",
            "title": "Patient Id"
          },
          "ingredients": {
            "items": {
              "$ref": "#/components/schemas/IngredientInput"
            },
            "type": "array",
            "title": "Ingredients"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "bolus_units_administered": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bolus Units Administered"
          }
        },
        "type": "object",
        "required": [
          "patient_id",
          "ingredients"
        ],
        "title": "LogMealRequest"
      },
      "MealLogResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "patient_id": {
            "type": "string",
            "format": "uuid",
            "title": "Patient Id"
          },
          "total_carbs_grams": {
            "type": "number",
            "title": "Total Carbs Grams"
          },
          "total_glycemic_load": {
            "type": "number",
            "title": "Total Glycemic Load"
          },
          "bolus_units_administered": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bolus Units Administered"
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "id",
          "patient_id",
          "total_carbs_grams",
          "total_glycemic_load"
        ],
        "title": "MealLogResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}